pft: Implementation of new PFT API. Policy set has to hold the data model
